Move SSLTransport from contrib to utils.
